requirements.txt
streamlit==1.39.0
pandas==2.2.2
openpyxl==3.1.5          # for Excel support
plotly==5.22.0
python-dotenv==1.0.1
dashscope==1.19.0        # Official Qwen SDK